requirements.txt
langchain==0.3.9
langchain_community==0.3.8
langchain_core==0.3.21
langchain_openai==0.2.10
langsmith==0.1.147
openai==1.55.2
PyPDF2==3.0.1
python-dotenv==1.0.1
python_docx==1.1.2
streamlit==1.33.0
streamlit_pdf_viewer==0.0.19
